# Fitter.App/API/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "Fitter API"
    },
    "paths": {
        "/api/comments/create": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "operationId": "CreateComment",
                "consumes": [
                    "application/json-patch+json",
                    "application/json",
                    "text/json",
                    "application/*+json"
                ],
                "produces": [],
                "parameters": [
                    {
                        "name": "comment",
                        "in": "body",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/CommentModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/comments/delete": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "operationId": "DeleteComment",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/comments/tagUsers": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "operationId": "CommentTagUsers",
                "consumes": [
                    "application/json-patch+json",
                    "application/json",
                    "text/json",
                    "application/*+json"
                ],
                "produces": [],
                "parameters": [
                    {
                        "name": "users",
                        "in": "body",
                        "required": false,
                        "schema": {
                            "uniqueItems": false,
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/UserDetailModel"
                            }
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/comments/getCommentsForPost": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "operationId": "GetCommentsForPost",
                "consumes": [],
                "produces": [
                    "text/plain",
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "uniqueItems": false,
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CommentModel"
                            }
                        }
                    }
                }
            }
        },
        "/api/comments/getTagsForComment": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "operationId": "GetTagsForComment",
                "consumes": [],
                "produces": [
                    "text/plain",
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "uniqueItems": false,
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/UserListModel"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "CommentModel": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/UserDetailModel"
                },
                "created": {
                    "format": "date-time",
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "post": {
                    "$ref": "#/definitions/PostModel"
                },
                "id": {
                    "format": "uuid",
                    "type": "string"
                }
            }
        },
        "UserDetailModel": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "id": {
                    "format": "uuid",
                    "type": "string"
                }
            }
        },
        "PostModel": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/UserDetailModel"
                },
                "created": {
                    "format": "date-time",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "team": {
                    "$ref": "#/definitions/TeamDetailModel"
                },
                "id": {
                    "format": "uuid",
                    "type": "string"
                }
            }
        },
        "TeamDetailModel": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "admin": {
                    "$ref": "#/definitions/UserDetailModel"
                },
                "id": {
                    "format": "uuid",
                    "type": "string"
                }
            }
        },
        "UserListModel": {
            "type": "object",
            "properties": {
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "format": "uuid",
                    "type": "string"
                }
            }
        },
...
    },
...
}